Create chipseq.nf in chipseq branch
this is a rough draft of a nextflow workflow for ChipSeq. It assumes a sample sheet, known Reference genome, info about Treatment-Control pairs, info about need to call peaks and if peaks are broad. Makes a  a Coverage bigWig for all samples and a Differential T-vs-C bigWig file if appropriate. Not dealing with replicates. 
Uses containers for bowtie2, samtools, deepTools, macs3.